test: move common gtest setup to header file
